122 Line (352 tokens) duplication in files: UMLActivityDiagram and UMLStateDiagram. Issue #19
