+++ src/argouml-app/src/org/argouml/uml/diagram/activity/ui/UMLActivityDiagram.java
@@ -61,14 +61,10 @@ import org.argouml.model.Model;
 import org.argouml.ui.CmdCreateNode;
 import org.argouml.uml.diagram.DiagramElement;
 import org.argouml.uml.diagram.DiagramSettings;
+import org.argouml.uml.diagram.UMLActionDiagram;
 import org.argouml.uml.diagram.UMLMutableGraphSupport;
 import org.argouml.uml.diagram.activity.ActivityDiagramGraphModel;
 import org.argouml.uml.diagram.state.StateDiagramGraphModel;
-import org.argouml.uml.diagram.state.ui.ActionCreatePseudostate;
-import org.argouml.uml.diagram.state.ui.ButtonActionNewCallEvent;
-import org.argouml.uml.diagram.state.ui.ButtonActionNewChangeEvent;
-import org.argouml.uml.diagram.state.ui.ButtonActionNewSignalEvent;
-import org.argouml.uml.diagram.state.ui.ButtonActionNewTimeEvent;
 import org.argouml.uml.diagram.state.ui.FigBranchState;
 import org.argouml.uml.diagram.state.ui.FigFinalState;
 import org.argouml.uml.diagram.state.ui.FigForkState;
@@ -77,23 +73,10 @@ import org.argouml.uml.diagram.state.ui.FigJoinState;
 import org.argouml.uml.diagram.state.ui.FigJunctionState;
 import org.argouml.uml.diagram.state.ui.FigStateVertex;
 import org.argouml.uml.diagram.static_structure.ui.FigComment;
-import org.argouml.uml.diagram.ui.ActionSetMode;
 import org.argouml.uml.diagram.ui.FigNodeModelElement;
 import org.argouml.uml.diagram.ui.RadioAction;
-import org.argouml.uml.diagram.ui.UMLDiagram;
-import org.argouml.uml.ui.behavior.common_behavior.ActionNewActionSequence;
-import org.argouml.uml.ui.behavior.common_behavior.ActionNewCallAction;
-import org.argouml.uml.ui.behavior.common_behavior.ActionNewCreateAction;
-import org.argouml.uml.ui.behavior.common_behavior.ActionNewDestroyAction;
-import org.argouml.uml.ui.behavior.common_behavior.ActionNewReturnAction;
-import org.argouml.uml.ui.behavior.common_behavior.ActionNewSendAction;
-import org.argouml.uml.ui.behavior.common_behavior.ActionNewTerminateAction;
-import org.argouml.uml.ui.behavior.common_behavior.ActionNewUninterpretedAction;
-import org.argouml.uml.ui.behavior.state_machines.ButtonActionNewGuard;
-import org.argouml.util.ToolBarUtility;
 import org.tigris.gef.base.LayerPerspective;
 import org.tigris.gef.base.LayerPerspectiveMutable;
-import org.tigris.gef.base.ModeCreatePolyEdge;
 import org.tigris.gef.graph.GraphModel;
 import org.tigris.gef.presentation.Fig;
 import org.tigris.gef.presentation.FigNode;
@@ -103,7 +86,7 @@ import org.tigris.gef.presentation.FigNode;
  *
  * TODO: Finish the work on subactivity states.
  */
-public class UMLActivityDiagram extends UMLDiagram implements ActivityDiagram {
+public class UMLActivityDiagram extends UMLActionDiagram implements ActivityDiagram {
 
     /**
      * Logger.
@@ -122,29 +105,10 @@ public class UMLActivityDiagram extends UMLDiagram implements ActivityDiagram {
     private Object theActivityGraph;
 
     private Action actionState;
-    private Action actionStartPseudoState;
-    private Action actionFinalPseudoState;
-    private Action actionJunctionPseudoState;
-    private Action actionForkPseudoState;
-    private Action actionJoinPseudoState;
-    private Action actionTransition;
     private Action actionObjectFlowState;
     private Action actionSwimlane;
     private Action actionCallState;
     private Action actionSubactivityState;
-    private Action actionCallEvent;
-    private Action actionChangeEvent;
-    private Action actionSignalEvent;
-    private Action actionTimeEvent;
-    private Action actionGuard;
-    private Action actionCallAction;
-    private Action actionCreateAction;
-    private Action actionDestroyAction;
-    private Action actionReturnAction;
-    private Action actionSendAction;
-    private Action actionTerminateAction;
-    private Action actionUninterpretedAction;
-    private Action actionActionSequence;
 
     /**
      * Constructor.
@@ -385,32 +349,6 @@ public class UMLActivityDiagram extends UMLDiagram implements ActivityDiagram {
         return actions;
     }
 
-    protected Object[] getTriggerActions() {
-        Object[] actions = {
-            getActionCallEvent(),
-            getActionChangeEvent(),
-            getActionSignalEvent(),
-            getActionTimeEvent(),
-        };
-        ToolBarUtility.manageDefault(actions, "diagram.activity.trigger");
-        return actions;
-    }
-
-    protected Object[] getEffectActions() {
-        Object[] actions = {
-            getActionCallAction(),
-            getActionCreateAction(),
-            getActionDestroyAction(),
-            getActionReturnAction(),
-            getActionSendAction(),
-            getActionTerminateAction(),
-            getActionUninterpretedAction(),
-            getActionActionSequence(),
-        };
-        ToolBarUtility.manageDefault(actions, "diagram.activity.effect");
-        return actions;
-    }
-
     /*
      * @see org.argouml.uml.diagram.ui.UMLDiagram#getLabelName()
      */
@@ -430,58 +368,6 @@ public class UMLActivityDiagram extends UMLDiagram implements ActivityDiagram {
                                 "button.new-callstate"));
         }
         return actionCallState;
-    }
-    /**
-     * @return Returns the actionFinalPseudoState.
-     */
-    protected Action getActionFinalPseudoState() {
-        if (actionFinalPseudoState == null) {
-            actionFinalPseudoState =
-                new RadioAction(
-                        new CmdCreateNode(
-                                Model.getMetaTypes().getFinalState(),
-                        	"button.new-finalstate"));
-        }
-        return actionFinalPseudoState;
-    }
-    /**
-     * @return Returns the actionForkPseudoState.
-     */
-    protected Action getActionForkPseudoState() {
-        if (actionForkPseudoState == null) {
-            actionForkPseudoState =
-                new RadioAction(
-                        new ActionCreatePseudostate(
-                                Model.getPseudostateKind().getFork(),
-                        	"button.new-fork"));
-        }
-        return actionForkPseudoState;
-    }
-    /**
-     * @return Returns the actionJoinPseudoState.
-     */
-    protected Action getActionJoinPseudoState() {
-        if (actionJoinPseudoState == null) {
-            actionJoinPseudoState =
-                new RadioAction(
-                        new ActionCreatePseudostate(
-                                Model.getPseudostateKind().getJoin(),
-                        	"button.new-join"));
-        }
-        return actionJoinPseudoState;
-    }
-    /**
-     * @return Returns the actionJunctionPseudoState.
-     */
-    protected Action getActionJunctionPseudoState() {
-        if (actionJunctionPseudoState == null) {
-            actionJunctionPseudoState =
-                new RadioAction(
-                        new ActionCreatePseudostate(
-                                Model.getPseudostateKind().getJunction(),
-                                "button.new-junction"));
-        }
-        return actionJunctionPseudoState;
     }
     /**
      * @return Returns the actionSwimlane.
@@ -507,28 +393,14 @@ public class UMLActivityDiagram extends UMLDiagram implements ActivityDiagram {
         return actionObjectFlowState;
     }
     /**
-     * @return Returns the actionStartPseudoState.
-     */
-    protected Action getActionStartPseudoState() {
-        if (actionStartPseudoState == null) {
-            actionStartPseudoState =
-                new RadioAction(
-                        new ActionCreatePseudostate(
-                                Model.getPseudostateKind().getInitial(),
-                                "button.new-initial"));
-        }
-        return actionStartPseudoState;
-    }
-    /**
      * @return Returns the actionState.
      */
     protected Action getActionState() {
         if (actionState == null) {
             actionState =
                 new RadioAction(
-                        new CmdCreateNode(
-                                Model.getMetaTypes().getActionState(),
-                        	"button.new-actionstate"));
+                        new CmdCreateNode(Model.getMetaTypes().getActionState(),
+                                "button.new-actionstate"));
         }
         return actionState;
     }
@@ -545,161 +417,12 @@ public class UMLActivityDiagram extends UMLDiagram implements ActivityDiagram {
         }
         return actionSubactivityState;
     }
-    /**
-     * @return Returns the actionTransition.
-     */
-    protected Action getActionTransition() {
-        if (actionTransition == null) {
-            actionTransition =
-                new RadioAction(
-                        new ActionSetMode(
-                                ModeCreatePolyEdge.class,
-                                "edgeClass",
-                                Model.getMetaTypes().getTransition(),
-                        "button.new-transition"));
-        }
-        return actionTransition;
-    }
-
-    /**
-     * @return Returns the actionCallEvent.
-     */
-    protected Action getActionCallEvent() {
-        if (actionCallEvent == null) {
-            actionCallEvent = new ButtonActionNewCallEvent();
-        }
-        return actionCallEvent;
-    }
-
-    /**
-     * @return Returns the actionCallEvent.
-     */
-    protected Action getActionChangeEvent() {
-        if (actionChangeEvent == null) {
-            actionChangeEvent = new ButtonActionNewChangeEvent();
-        }
-        return actionChangeEvent;
-    }
-
-    /**
-     * @return Returns the actionCallEvent.
-     */
-    protected Action getActionSignalEvent() {
-        if (actionSignalEvent == null) {
-            actionSignalEvent = new ButtonActionNewSignalEvent();
-        }
-        return actionSignalEvent;
-    }
-
-    /**
-     * @return Returns the actionCallEvent.
-     */
-    protected Action getActionTimeEvent() {
-        if (actionTimeEvent == null) {
-            actionTimeEvent = new ButtonActionNewTimeEvent();
-        }
-        return actionTimeEvent;
-    }
-
-    protected Action getActionGuard() {
-        if (actionGuard == null) {
-            actionGuard = new ButtonActionNewGuard();
-        }
-        return actionGuard;
-    }
-
-    protected Action getActionCallAction() {
-        if (actionCallAction == null) {
-            actionCallAction = ActionNewCallAction.getButtonInstance();
-        }
-        return actionCallAction;
-    }
-
-    protected Action getActionCreateAction() {
-        if (actionCreateAction == null) {
-            actionCreateAction = ActionNewCreateAction.getButtonInstance();
-        }
-        return actionCreateAction;
-    }
-
-    protected Action getActionDestroyAction() {
-        if (actionDestroyAction == null) {
-            actionDestroyAction = ActionNewDestroyAction.getButtonInstance();
-        }
-        return actionDestroyAction;
-    }
-
-    protected Action getActionReturnAction() {
-        if (actionReturnAction == null) {
-            actionReturnAction = ActionNewReturnAction.getButtonInstance();
-        }
-        return actionReturnAction;
-    }
-
-    protected Action getActionSendAction() {
-        if (actionSendAction == null) {
-            actionSendAction = ActionNewSendAction.getButtonInstance();
-        }
-        return actionSendAction;
-    }
-
-    protected Action getActionTerminateAction() {
-        if (actionTerminateAction == null) {
-            actionTerminateAction =
-                ActionNewTerminateAction.getButtonInstance();
-        }
-        return actionTerminateAction;
-    }
-
-    protected Action getActionUninterpretedAction() {
-        if (actionUninterpretedAction == null) {
-            actionUninterpretedAction =
-                ActionNewUninterpretedAction.getButtonInstance();
-        }
-        return actionUninterpretedAction;
-    }
-
-    protected Action getActionActionSequence() {
-        if (actionActionSequence == null) {
-            actionActionSequence =
-                ActionNewActionSequence.getButtonInstance();
-        }
-        return actionActionSequence;
-    }
 
     /*
      * @see org.argouml.uml.diagram.ui.UMLDiagram#getDependentElement()
      */
     public Object getDependentElement() {
         return getStateMachine(); /* The ActivityGraph. */
-    }
-
-    /*
-     * @see org.argouml.uml.diagram.ui.UMLDiagram#isRelocationAllowed(java.lang.Object)
-     */
-    public boolean isRelocationAllowed(Object base) {
-        return false;
-        /* TODO: We may return the following when the
-         * relocate() has been implemented.
-         */
-//      Model.getActivityGraphsHelper()
-//      .isAddingActivityGraphAllowed(base);
-    }
-
-    @SuppressWarnings("unchecked")
-    public Collection getRelocationCandidates(Object root) {
-        /* TODO: We may return something useful when the
-         * relocate() has been implemented. */
-        Collection c =  new HashSet();
-        c.add(getOwner());
-        return c;
-    }
-
-    /*
-     * @see org.argouml.uml.diagram.ui.UMLDiagram#relocate(java.lang.Object)
-     */
-    public boolean relocate(Object base) {
-        return false;
     }
 
     /**
